backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.9
psycopg2-binary==2.9.6
redis==4.5.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiohttp==3.8.4
reportlab==3.6.12  # For PDF generation